Give bounds of field lattice and vacuum values
Rather than assuming that the field is at the vacua at the edges of
its bounding box, we can take a bounding box and vacuum values and
work out the steps. This is important for when adding a second
field, as only the single-field case can be always set up as having
vacuum values bounding the lattice.