Extend the README with observations
After revisiting the literature, I discovered that I could have
left the fixed end qubits of each word as fictitious. I also
noticed that I did not correctly turn the integral into a sum,
because I neglected the factor of the lattice spacing per term.
Fortunately, the lattice spacing is constant in the cases in the
program and does not affect the minimization. The final value of
the minimized objective function is already a mess because of the
extra terms enforcing ICDW and the bubble center and edge being at
the true vacuum and false vacuum respectively, so it is really not
important to correct. I feel that it is sufficient to just make
these observations in the README.